tests: better separation of unit tests and snapshot tests
