+++ tests/snapshot_tests/load_preprocess_test.py
@@ -1,9 +1,6 @@
-import pytest
 from rex_xai.explanation import (
     load_and_preprocess_data,
-    predict_target,
     try_preprocess,
-    validate_args,
 )
 
 
@@ -31,14 +28,6 @@ def test_preprocess_spectral_mask_on_image_returns_warning(
     assert data == snapshot
 
 
-def test_preprocess_nii_notimplemented(args, model_shape, cpu_device, caplog):
-    args.path = "tests/test_data/dog.nii"
-    data = try_preprocess(args, model_shape, device=cpu_device)
-
-    assert data == NotImplemented
-    assert caplog.records[0].msg == "we do not (yet) handle nifti files generically"
-
-
 def test_preprocess_npy(args, model_shape, cpu_device, snapshot, caplog):
     args.path = "tests/test_data/DoublePeakClass 0 Mean.npy"
     try_preprocess(args, model_shape, device=cpu_device)
@@ -48,16 +37,3 @@ def test_preprocess_npy(args, model_shape, cpu_device, snapshot, caplog):
     args.mode = "tabular"
     data = try_preprocess(args, model_shape, device=cpu_device)
     assert data == snapshot
-
-
-def test_predict_target(data, prediction_func):
-    target = predict_target(data, prediction_func)
-
-    assert target.classification == 207
-    assert target.confidence == pytest.approx(0.253237, abs=2.5e-6)
-
-
-def test_validate_args(args):
-    args.path = None  #  type: ignore
-    with pytest.raises(FileNotFoundError):
-        validate_args(args)
