+++ tests/snapshot_tests/spectral_test.py
